handling aftermath of moving multiple files
