+++ src/Pages/Collector/InputComponents/Experience.js
@@ -1,9 +1,9 @@
 import { Button, Stack, TextField } from "@mui/material";
 import { Box } from "@mui/system";
-import { addExperience, removeExperience } from "../../Redux/Actions/index";
+import { addExperience, removeExperience } from "../../../Redux/Actions/index";
 import { connect } from "react-redux";
 import { useState } from "react";
-import { EMPTY_EXPERIENCE } from "../../helper/InitialState";
+import { EMPTY_EXPERIENCE } from "../../../helper/InitialState";
 const AddExperience = (props) => {
   const isNotEmpty = (props.idx !== -1);
   const data = isNotEmpty ? props.experiences[props.idx] : EMPTY_EXPERIENCE;
