+++ src/Pages/Collector/InputComponents/Project.js
@@ -1,9 +1,9 @@
 import { Button, Stack, TextField } from "@mui/material";
 import { Box } from "@mui/system";
-import { addProject, removeProject } from "../../Redux/Actions/index";
+import { addProject, removeProject } from "../../../Redux/Actions/index";
 import { connect } from "react-redux";
 import { useState } from "react";
-import { EMPTY_PROJECT } from "../../helper/InitialState";
+import { EMPTY_PROJECT } from "../../../helper/InitialState";
 const AddProject = (props) => {
   const isNotEmpty = (props.idx !== -1);
   const data = isNotEmpty ? props.projects[props.idx] : EMPTY_PROJECT;
